Flytt hvor man redirecter
